fix: handler 변경 및 타입 에러 처리
